feat: uv unwrap on partition 2/3

+++ src/terrainlib/mesh/merging/VertexId.h
@@ -6,8 +6,8 @@ namespace mesh::merging {
 
 // Identifies a specific vertex within a particular mesh.
 struct VertexId {
-    size_t mesh_index;
-    size_t vertex_index;
+    uint32_t mesh_index;
+    uint32_t vertex_index;
 
     auto operator<=>(const VertexId &) const = default;
 };
